Using URI instead of String Moving models to dedicated folder
